Fixes #19229: Deadlock at start when loading new licence
