--- branding/src/main/resources/branding/toserve/toasty.css
@@ -1,139 +1,0 @@
-.toasty-container {
-    padding: 1em;
-    border-radius: 5px;
-    cursor: pointer;
-    -webkit-box-shadow: 0 5px 5px -5px rgba(0, 0, 0, .5);
-    box-shadow: 0 5px 5px -5px rgba(0, 0, 0, .5);
-    color: #fff;
-    font-size: 13px
-}
-
-.toasty-title {
-    font-size: 1em;
-    margin: 0
-}
-
-.toasty-message {
-    font-size: .9em;
-    margin-top: .5em;
-    margin-bottom: 0
-}
-
-.toasty-success {
-    background-color: rgba(155, 200, 50, .9);
-    border: 1px solid #a4c74e
-;}
-
-.toasty-warning {
-    background-color: #ff851b
-}
-
-.toasty-error {
-    background-color: rgba(200, 0, 4, .9);
-    border: 1px solid #c3000e;
-}
-
-.animated {
-    -webkit-animation-duration: 1s;
-    animation-duration: 1s;
-    -webkit-animation-fill-mode: both;
-    animation-fill-mode: both
-}
-
-@-webkit-keyframes bounceInRight {
-    0%,
-    60%,
-    75%,
-    90%,
-    to {
-        -webkit-animation-timing-function: cubic-bezier(.215, .61, .355, 1);
-        animation-timing-function: cubic-bezier(.215, .61, .355, 1)
-    }
-    0% {
-        opacity: 0;
-        -webkit-transform: translate3d(3000px, 0, 0);
-        transform: translate3d(3000px, 0, 0)
-    }
-    60% {
-        opacity: 1;
-        -webkit-transform: translate3d(-25px, 0, 0);
-        transform: translate3d(-25px, 0, 0)
-    }
-    75% {
-        -webkit-transform: translate3d(10px, 0, 0);
-        transform: translate3d(10px, 0, 0)
-    }
-    90% {
-        -webkit-transform: translate3d(-5px, 0, 0);
-        transform: translate3d(-5px, 0, 0)
-    }
-    to {
-        -webkit-transform: none;
-        transform: none
-    }
-}
-
-@keyframes bounceInRight {
-    0%,
-    60%,
-    75%,
-    90%,
-    to {
-        -webkit-animation-timing-function: cubic-bezier(.215, .61, .355, 1);
-        animation-timing-function: cubic-bezier(.215, .61, .355, 1)
-    }
-    0% {
-        opacity: 0;
-        -webkit-transform: translate3d(3000px, 0, 0);
-        transform: translate3d(3000px, 0, 0)
-    }
-    60% {
-        opacity: 1;
-        -webkit-transform: translate3d(-25px, 0, 0);
-        transform: translate3d(-25px, 0, 0)
-    }
-    75% {
-        -webkit-transform: translate3d(10px, 0, 0);
-        transform: translate3d(10px, 0, 0)
-    }
-    90% {
-        -webkit-transform: translate3d(-5px, 0, 0);
-        transform: translate3d(-5px, 0, 0)
-    }
-    to {
-        -webkit-transform: none;
-        transform: none
-    }
-}
-
-.bounceInRight {
-    -webkit-animation-name: bounceInRight;
-    animation-name: bounceInRight
-}
-
-@-webkit-keyframes fadeOutRightBig {
-    0% {
-        opacity: 1
-    }
-    to {
-        opacity: 0;
-        -webkit-transform: translate3d(2000px, 0, 0);
-        transform: translate3d(2000px, 0, 0)
-    }
-}
-
-@keyframes fadeOutRightBig {
-    0% {
-        opacity: 1
-    }
-    to {
-        opacity: 0;
-        -webkit-transform: translate3d(2000px, 0, 0);
-        transform: translate3d(2000px, 0, 0)
-    }
-}
-
-.fadeOutRightBig {
-    -webkit-animation-name: fadeOutRightBig;
-    animation-name: fadeOutRightBig
-}
